Patch from Scott McKellar:
1. I added the const qualifier to a number of function parameters.

2. I moved the prototype for _oilsEventParseEvents() from the header
into the implementation file, and made the function static.  No other
source file calls it, nor should it.

3. I removed an extra leading underscore from each of _oilsEventEvents
and _oilsEventDescriptions, and made them static.

3. I removed an unhelpful cast from a call to safe_malloc().

4. I made sure to initialize every member of a new oilsEvent.

5. In several spots where we update pointer members of an oilsEvent,
I preceded the update with a free, in order to avoid potential
memory leaks.

6. I replaced calls to oilsEventSetPermission() and
oilsEventSetPayload() with the equivalent inline code.

7. In oilsEventFree(), the original code would free the json member
or the payload member but not both.  We now free both.  We also
free the event member, which we didn't do before.