Patch to better handle migration of settings on the filesystem.  Thanks Jeff!
===
Greetings-

I found that when using open-ils.write_in_user_chrome_directory set to
true, the staff client would store local settings in the application
chrome directory, but the next time the client was run, it would
inappropriately attempt to migrate the settings from the application
chrome directory to the user chrome directory.

If you allowed the settings to be migrated, on the third run, things would
get even more confused.

The attached patch causes the migration check in main_init() to respect
the open-ils.write_in_user_chrome_directory preference, and not migrate
the local settings out of the application chrome directory if the
preference dictates that they should remain there.

DCO inline below, patch is attached. Feedback welcome!

-jeff

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
   have the right to submit it under the open source license
   indicated in the file; or

(b) The contribution is based upon previous work that, to the best
   of my knowledge, is covered under an appropriate open source
   license and I have the right under that license to submit that
   work with modifications, whether created in whole or in part
   by me, under the same open source license (unless I am
   permitted to submit under a different license), as indicated
   in the file; or

(c) The contribution was provided directly to me by some other
   person who certified (a), (b) or (c) and I have not modified
   it.

(d) I understand and agree that this project and the contribution
   are public and that a record of the contribution (including all
   personal information I submit with it, including my sign-off) is
   maintained indefinitely and may be redistributed consistent with
   this project or the open source license(s) involved.

Signed-off-by: Jeff Godin <jeff@tcnet.org>

--
Jeff Godin <jeff@tcnet.org>
Traverse Area District Library