Refactor formatting code to properly handle chronology information stored in the enumeration fields (well, I think it's handling it properly; it hasn't been tested yet for that case).
Also refactored basic next/prediction code to handle the same
situation. Basically this is all minor stuff that needs to be
done, but allows me to avoid working on actually taking a stab at
predicting the date of the next issue, but that's pretty much all
that's left to do for this iteration of the code.