Eliminated all calls to oilsIDLFindPath() in this file.
This change eliminates the overhead of parsing path
strings, including the malloc and free that such
parsing entails.

In addition: in most cases we had been calling
oilsIDLFindPath() twice, once to verify that the class
existed, and again to find whatever piece of it we
wanted.  This duplication of effort is now gone.
-This line, and those below, will be ignored--

M    Open-ILS/src/c-apps/oils_idl-core.c