Plugged some massive memory leaks.
In one case we were leaking an osrfHash whenever we didn't use it.
I rearranged matters so that we won't create the osrfHash unless
we're going to use it.

The other leaks involved various libxml2 functions that return
allocated strings that need to be freed.  In many of those cases
we had been stdruping the strings and leaking the originals.
Eliminating all those strdups should boost performance a bit,
and reduce the memory footprint.