1. In osrfAppChildInit(): reduce the scope of the attr variable.  Don't call dbi_result_get_field_attribs() unless you're going to use the result.
2. In dbi_result_get_field_attribs(): Eliminate the round trip from string to
number and back to string; just use the string that's already available.

If the string is in a JSON_NUMBER, it's guaranteed to be numeric.  If the
string is in a JSON_STRING, it might not be numeric, and if it isn't, then
PostgreSQL will choke on it.  The old code would silently convert it to zero,
or possibly to some other unintended value.