Patch from Warren Layton implementing RefWorks export via SuperCat:
"The patch a new transform, MARC21slim2MARCtxt.xsl, which converts MARC
into the MARC-ish plain text format accepted by RefWorks (see [1]).
The patch also adds a new feed to SuperCat, called "marctxt"."

As well as a reworked patch for the default opac skin to expose this
functionality as a link next to the "Place a Hold" link.