Fix SQL translation script and associated unit tests
It looks like polib became stricter about forcing the line number portion of 
filename:line occurrences in the PO to actually be an integer, with the result
of simply concatenating the text "line numbers" that we had been using to
store ID values to the table names.

Now we're using specially formatted translator comments to hold the ID values,
and using the filename:line occurence entries to actually hold the source filename
and line number for occurrences. Fancy that!