Several tweaks to oils_cstore.c:
1. Added a utility function single_hash() to create simple WHERE
clauses instead of calling jsonParseStringFmt().  This change
avoids the overhead of creating and then parsing a JSON string.

2. In SELECT(), processing an ORDER BY clause: traverse a
JSON_ARRAY with an index instead of an iterator.  This change
avoids some mallocs and frees -- and also plugs a memory leak,
since we weren't freeing the iterator.

3. Also in SELECT(), processing an ORDER BY clause: free the
iterator that we use for traversing the hash of classes.