Tightened the rules for defining SELECT clauses for a JSON query.
With two exceptions as noted below, the SELECT list for every class
must be encoded as an array.  Anything else is an error.

The exceptions: null and "*".  For the core class only, these
encodings request a default SELECT list.

For the core class: an empty array requests a default SELECT list.

For a non-core class: an empty array results in an error message,
but not an error.