Alternate Item Status View
This is a work in process and I expect a lot of churn; I promise to I18N-ize it before it leaves trunk.  Some stuff like Rolling Counter is more mock-up than anything; I have no idea what a Rolling Counter is.  :D

The main notion here is that we want a quick toggle in Item Status for a denser view of information for a single item.  Something less like the current Actions -> Show Item Details and more like the MARC View button in the Z39.50 
interface.  Note, you can scan in new barcodes while remaining in this view.   In general, I want to move away from pop-up dialogs and interfaces in any case, and this sort of deck swapping is probably the future.

Feedback welcome.