Make building & installing the core components of Evergreen optional.
For now, this will be of most interest to those interested in building only
the staff client, say, on platforms where many of the dependencies of the
other components of Evergreen are not readily available. If you disable the
core components using configure, then dependency checking for those core
components is skipped and the staff client will be build and installed to
the desired location.

For example:

./autogen.sh
./configure --prefix=/openils --disable-core --disable-apache-modules
sudo make install

... will result in the Evergreen Web files, reporter, and staff client
being built and their respective server components installed in
/openils/var/web/...