In XUL, if you do something like this:
   <label control="foo" value="My Label">
   <textbox id="foo" />

Then clicking on the label will jump focus to the textbox.  Replacing the onclick handlers in the MARC editor with that way of doing it.

Also having each fixed field auto-select all the text on focus, so whether a user clicks on the label or directly into the textbox, or even tabs into the textbox, the text will be selected so that the user can immediately type a 
new replacement value.