Support ui.circ.show_billing_tab_on_bills org setting.
Some refactoring to eliminate a redundant network call and to have the stop sign page trump the Show Billing Tab on Bills behavior.

This whole display/summary/sub-interface setup needs some TLC, and I want to redo the stop sign page so that you can spawn it quickly and just feed it asynchronous data as needed.