Toward a new look for XUL Record Summary.  bib_brief.xul modified for existing instances of Record Summary, but the alternate Item Details interface is doing it differently.
The old Record Summary is a dedicated XUL file that typically lives in an iframe.  It would fetch its own data if fed some ID's either via xulG or URL parameters.

The new way (which is really an old way that was tried before and is used with bill summaries) involves a XUL overlay, and a render function matching the overlay name.  A caller can feed it already fetched objects or have it fetch 
its own objects given ID's.

To do this, 

Include these in your XUL file:  
	<?xul-overlay href="/xul/server/cat/bib_brief_overlay.xul"?>
	<grid id="bib_brief_grid"/>

This pulls in the overlay, which merges with the <grid/> in your XUL and loads bib_brief_overlay.js.

In your javascript, call bib_brief_overlay():

	bib_brief_overlay( { 'mvr' : details.mvr, 'acp' : details.copy } );

Different options include mvr, mvr_id, bre, bre_id, and acp.  This will render values into the grid.

Since this is an overlay, you can mix-in your own elements.  The new Item Details is doing this:

	<grid id="bib_brief_grid">
	    <rows id="bib_brief_grid_rows">
		<row>
		    <label value="&staff.circ.alternate_copy_summary.Call_Number.label;" accesskey="&staff.circ.alternate_copy_summary.call_number.accesskey;" control="call_number"/>
		    <textbox name="call_number" readonly="true" context="clipboard"/>
		</row>
	    </rows>
	</grid>

This is adding a custom/local row to the bottom of the Record Summary grid.