Cleaning up some syntax thinkos from earlier commit which added support for hold target weighting and max hold targeting loops.
Added closed-date check condition such that if a library is closed both
now and at the next expected check time the location is ignored for
targetting.  This differs from before in that we were only checking to
see if the library was currently closed.  For 24h retargetting, this is
nominally not a change, but for longer-that-one-day retargetting intervals
this will allow a library that is open at the beginning /or/ the end of the
targetting interval to have a chance at capturing the requested copy.