Fix minor typos for plural versus singular strings, and change a few identifiters to protect against accidental column collisions.
Meat of this changeset is to allow manipulation of the mint_condition fields on items and holds.  The desired functionality is a way to let holds opt for pristine/complete/or-just-plain-better copies 
or not, with the idea being that an item missing pieces or slightly damaged could be marked as mint_condition = false.  I've re-cast mint condition as Quality/Copy Quality as far as labels go, with 
Good and Mediocre/Any Copy instead of True or False.  I'm open to better suggestions.  There's some dissonance here with the Damaged copy status.