Batch barcode editing and copy (item attribute) editor tweaks.  A few things here, and I apologize for not cherry-picking into separate commits:
* Modified the volume/copy creator to serve double-duty as an existing copy rebarcoder UI.  This also gets us closer to being able to shove in existing pre-cats as an alternative cataloging workflow to transfering pre-cats to 
targeted volumes.

* Tested this out as a Replace Barcode replacement in Holdings Maintenance, but the real motivator was for the Update Barcodes link for lineitems in acquisitions.  

* The volume/copy interface now has two options, the new one bypasses the item attribute editor and creates or modifies the items instantly.

* Button labeling tweaks in the item attribute editor.

* Allow folks to change the item status for new copies.  I can't recall the rationale for disallowing this in the first place, but I've heard enough arguments asking for us to allow this.  If need be, we can wrap it in a perm if 
folks do want to lock it down .  We may also want to implement an org setting at some point for configuring the default status.

TODO: Replace "Replace Barcode" with this in Item Status UI

TODO: Allow batch call number editing.  There's an existing UI for this, but we need something in the context of acquisitions for taking care of those autogenerated call numbers and barcodes in one shot.