Support multiple languages in the lookup for event descriptions.
Capture the current locale, which reflects the message most
recently received, and confine the lookup to messages in that
language.  If you don't find a description, and the language is
different from the default language (hard-coded as "en-US"),
try again with the default language.

Also: eliminate a needless memset().

M    Open-ILS/src/c-apps/oils_event.c