Make GetOpts::Long and Cronscript get along
So the old Cronscript approach was to have variables declared with
the default values assigned and pass a hash of keys => var_ref, where
the var_ref's pointed to the variables.  This didn't work because of
Cronscript trying to build the clean hashes and incorporate structure
into the object.

So instead now we rely on the MyGetOptions return
value (hashref or hash), essentially the opts_clean hash.  And we
populate the local variables afterwards.  The default values are passed
in the hash now.  This works.