Acq: optionally mark POs as prepayment_required; show/confirm it where needed
The PO creation dialog box now has a checkbox for "prepayment required." This
will be checked or unchecked automatically when choosing a vendor from the
above dropdown based on whether that vendor's "prepayment required" value is
true.  The user can override the default value for a vendor, but will face a
confirmation dialog about it.

The PO viewing interface shows whether or not a PO is marked "prepayment
required," and if it is, when the user wishes to activate the PO, there will
be another confirmation dialog reminding the user to deal with payment.