Patch from Joe Atzberger:  Cleanup subs in Actor
For the first, @xact was never used.
The 3 part conditional only had two different outcomes.
There is still a problem with user_transactions reproducing a lot
of the functionality of user_transaction_history, which it already
calls to do the real query.  But it doesn't reproduce all of it.
Recommend relying on user_transaction_history as much as possible,
but that kind of work is out of scope for this review.

For the other, remove unused var and unnecessary array.
You don't need to build an array when all you do at the end is count its elements.

Signed-off-by: Joe Atzberger <atz@esilibrary.com>