Adds a xulG.get_new_session() convenience function that can be called by embedded dojo interfaces, etc.
Side effect: Sets a session cookie and updates the global OpenILS.data store with the newly acquired authtoken.

Return Value: True if a new authtoken was acquired, otherwise False.

Optional Param: Object with optional keys 'callback' and 'operator_change'.

If 'operator_change' is passed in with a value of true, then the dialog will resemble the Admin->Operator Change dialog, but it is not yet completely implemented (doesn't preserve the existing authtoken).  Otherwise, we get the normal Session Expired dialog.

Callback may be a function that resembles this:

    function( success, authtoken, authtime ) {
        return success; // whatever gets returned here will also be returned by xulG.get_new_session()
    }