Wire up the Clear Shelf functionality into the Hold Shelf Browser.
The button for this enables when the View Expired On-Shelf Holds checkbox is activated.  The Clear Shelf process will cancel these holds, optionally set the status for the affected items according to an org unit setting,  and report back the needed actions for the affected items.

TODO: the ML clear_shelf_process method needs fixing where it's jumping to storage within a transaction