XBL for a <help> widget.  It will try to load help content from various static and dynamic URL's, stopping at the first one that it finds.  If for example, I were to add this to /xul/server/patron/display.xul:
    <help id="bar" src="foo.html" />

We would get a "Help" button that if clicked would then it would test these locations:

    foo.html (if @foo not set, then custom_help.html)
    /xul/server/patron/display.xul.help.html
    /xul/server/patron/help.html
    /xul/server/help.html
    /xul/help.html
    /help.html

and since @id is defined, it would also tack on ?id=bar (for log tracking and/or dynamic use by the content).

It will open the location in a non-modal pop-up window.  Navigation buttons not included.

MAYBE TODO:
    Add an @anchor for dynamically appending a hash component to each URL tested.
    Implement a browser with Back/Forward/Search/Index functionality.