Reworking the <help> widget to add more hooks for localized/custom content, and a @pathname attribute for overriding the location.pathname value used to derive URL's for content.
Current behavior:

This widget will try to load help content from various static and dynamic URL's, stopping at the
first one that it finds.  Given an example location.href of '/xul/server/patron/display.xul' and
a @src of 'foo.html', it will check these URL's in this order:

    foo.html
    /xul/server/patron/display.xul.custom_help.html
    /xul/server/patron/display.xul.help.html
    /xul/server/patron/custom_help.html
    /xul/server/patron/help.html
    /xul/server/custom_help.html
    /xul/server/help.html
    /xul/custom_help.html
    /xul/help.html
    /custom_help.html
    /help.html

If @pathname is specified, it will override the value from location.href

The idea is that *custom_help.html files will never be stock.