Acq: add search-by-file-of-terms to unified search
This feature was present in the now deprecated lineitem search interface.
It works, but it is severely limited at the moment in the number of search
terms you can use, because search queries are made into URIs before the
search actually happens (in order to make backing up to old search results
possible), and at a certain point, very large URIs become infeasible.

Some adjustments to how search history is made naviagable may be in order.