Prevent possible character corruption or fatal errors in SRU / Z39.50
All MARC records in the Evergreen database are Unicode-encoded. However,
if the LDR09 hasn't been updated to indicate this, when we invoke
MARC::File::XML->as_xml_record it automatically runs a MARC8-to-UTF8 conversion
over the data - potentially corrupting the data, or resulting in fatal
errors. When holdings are added to records in SRU / Z39.50, we invoke
this code path.

The solution here is to simply manually set the leader 09 position to 'a'
before invoking MARC::File::XML->as_xml_record.